refactor: move network thunks
