+++ src/app/features/network/index.ts
@@ -1,2 +1,3 @@
 export * from './reducer';
 export * from './selectors';
+export * from './thunks';
