+++ src/app/features/network/reducer.ts
@@ -1,7 +1,7 @@
 import { createSlice } from '@reduxjs/toolkit';
 import { SupportedChainId } from '@voltz-protocol/v1-sdk';
 
-import { setChainIdThunk } from './thunks';
+import { setChainIdThunk } from './thunks/setChainIdThunk';
 
 type SliceState = {
   chainId: SupportedChainId | null;
