+++ src/app/features/network/thunks/index.ts
@@ -1,0 +1,1 @@
+export * from './setChainIdThunk';
