+++ src/ui/components/VoltzPage/TopPanel/ChainSelector/index.tsx
@@ -2,8 +2,11 @@ import { SupportedChainId } from '@voltz-protocol/v1-sdk';
 import { ChainSelector as BrokoliChainSelector } from 'brokoli-ui';
 import React, { useCallback, useEffect, useMemo } from 'react';
 
-import { selectChainChangeState, selectChainId } from '../../../../../app/features/network';
-import { setChainIdThunk } from '../../../../../app/features/network/thunks';
+import {
+  selectChainChangeState,
+  selectChainId,
+  setChainIdThunk,
+} from '../../../../../app/features/network';
 import { useAppDispatch, useAppSelector } from '../../../../../app/hooks';
 import { useWallet } from '../../../../../hooks/useWallet';
 import { setChainId } from '../../../../../utilities/network/chain-store';
